[add-booking] [feature] *add booking logic

+++ src/main/java/ru/practicum/shareit/item/ItemDto.java
@@ -21,5 +21,5 @@ public class ItemDto {
     @NonNull
     Long userId;
     @Nullable
-    Integer requestId;
+    Long requestId;
 }
